omg finish! did research, its because the button created after page load does not have an even listener attached to them.  A pretty common error is to have an init event on document.ready that attaches listeners to buttons like this:
$('.button').on('click',someAction);
That will work for buttons already on the page, but if buttons are added afterwards in your Ajax request the actions won’t be attached to your new buttons. Instead, in your document.ready event, attach the listeners to the document and they’ll work even on elements that are added after the page has finished loading.

$(document).on('click','.button',someAction);